requirements.txt
streamlit 
openai
python-dotenv
langchain
faiss-cpu
streamlit-extras
pypdf